Define foreign keys for relational sources

+++ pakyow-data/spec/spec_helper.rb
@@ -21,7 +21,16 @@ RSpec.configure do |config|
       connection.adapter.connection.nil?
     }.each do |connection|
       connection.adapter.connection.tables.each do |table|
-        connection.adapter.connection.run "DROP TABLE #{table}"
+        case connection.opts[:adapter]
+        when "sqlite"
+          connection.adapter.connection.run "PRAGMA foreign_keys = off"
+          connection.adapter.connection.run "DROP TABLE #{table}"
+        when "mysql2"
+          connection.adapter.connection.run "SET FOREIGN_KEY_CHECKS = 0"
+          connection.adapter.connection.run "DROP TABLE #{table}"
+        else
+          connection.adapter.connection.run "DROP TABLE #{table} CASCADE"
+        end
       end
     end
